Update swift to point to project entry
Swift is a project that is housed by the OpenStack Foundation, not by
the company SwiftStack. SwiftStack is also a fine group of humans, but
the https://github.com/cncf/landscape#new-entries rule says:

"We are generally not including commercial versions of open source software.
The exception is that we are showing all Certified Kubernetes and non-Certified
Kubernetes implementations."

The Swift shipped by SwiftStack is just the OpenStack Swift, so just
reference the foundation-related links rather than the company specific
ones.