Add Codefresh products to appropriate categories
The Codefresh registry launched less than 6 months ago and already stores 3-4 million images and is used by thousands on both the free and paid tiers. 

Codefresh Helm is a commercial implementation of the open source project by the same name and layers on a lot of functionality for testing configurations, managing releases between environments, etc.