Add Porter to CNCF landscape (#1737)
* Add porter.svg

* Add Porter to CNCF landscape

Porter is an open source load balancer designed for bare metal Kubernetes clusters. It's implemented by the physical switch and uses BGP and ECMP to achieve the best performance and high availability. Porter allows users to create Kubernetes services of type "LoadBalancer" in bare metal clusters, which makes you have a consistent experience with the cloud.

Porter is driven by the community, and there are three companies that contributed to this project.

* replace the Porter logo with a high resolution svg

* Delete porter.svg

* Rename Porter.svg to porter.svg